requirements.txt
# ─── Framework ──────────────────────────────────────────────────────────────────────
databricks-mlops-stack==1.5.0.2
# ─── Core ──────────────────────────────────────────────────────────────────────
typing_extensions>=4.14.1
pyspark
pytz~=2022.2.1
pillow>=10.0.1
Jinja2>=3.1.4
PyYAML
pytest
# ML stack
imbalanced-learn==0.14.0
lightgbm==4.3.0
matplotlib==3.10.8
mlflow==3.0.1
numpy==1.26.4
optuna==3.6.0
optuna-integration==3.6.0
pandas==2.3.3
ray[all]==2.37.0
scikit-learn==1.6.1
scipy==1.16.3
xgboost==2.1.4
# Azure
azure-keyvault==4.2.0
jsonschema






















